Added creatures panel and ability to select current creature to add. Now only adds with primary mouse button click on canvas.

+++ src/main/java/CreatureSimulator/creatures/SimpleCreature.java
@@ -6,7 +6,7 @@ import javafx.scene.paint.Color;
 
 public class SimpleCreature extends Creature {
 
-	public SimpleCreature(int x, int y) {
+	public SimpleCreature(double x, double y) {
 		super(x, y, WIDTH, WIDTH);
 	}
